requirements.txt
# OSMOTROFIA - Dependencias

# Sistema y monitoreo
psutil>=5.9.0

# Backend Web (REQUERIDO para v2.0)
flask>=3.0.0
flask-cors>=4.0.0

# Gmail API (opcional)
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0

# Visualización 2D - Legacy (opcional)
matplotlib>=3.5.0
numpy>=1.21.0
